client/TCPChatClient/requirements.txt
Eel
python-decouple
bcrypt
cryptography
